+++ sdpb/util/query.py
@@ -1,4 +1,3 @@
-import dateutil.parser
 import logging
 from flask import request
 from sqlalchemy import func
@@ -10,40 +9,6 @@ from sdpb.timing import log_timing
 logger = logging.getLogger("sdpb")
 
 
-def dict_from_row(row):
-    """Return a dict version of a SQLAlchemy result row"""
-    return dict(zip(row.keys(), row))
-
-
-def dicts_from_rows(rows):
-    """
-    Return a list of dicts constructed from a list of SQLAlchemy result rows
-    """
-    return [dict_from_row(row) for row in rows]
-
-
-def date_rep(date):
-    return date.isoformat() if date else None
-
-
-def float_rep(x):
-    return float(x) if x is not None else None
-
-
-def parse_date(s):
-    if s is None:
-        return None
-    return dateutil.parser.parse(s)
-
-
-def obs_stats_rep(obs_stats):
-    return {
-        "min_obs_time": date_rep(obs_stats.min_obs_time),
-        "max_obs_time": date_rep(obs_stats.max_obs_time),
-        # "obs_count": int(obs_stats.obs_count),
-    }
-
-
 def set_logger_level_from_qp(a_logger):
     """Set logger level from query parameter `debug`."""
     try:
@@ -51,9 +16,6 @@ def set_logger_level_from_qp(a_logger):
             a_logger.setLevel(logging.DEBUG)
     except:
         pass
-
-
-# TODO: Move these into a different util module
 
 
 def get_all_histories_etc_by_station(session):
@@ -67,6 +29,7 @@ def get_all_histories_etc_by_station(session):
                 StationObservationStats,
                 StationObservationStats.history_id == History.id,
             )
+            # TODO: Is ordering necessary? Probably not.
             .order_by(History.station_id.asc(), History.id)
             .all()
         )
@@ -101,14 +64,14 @@ def get_all_vars_by_hx(session, group_in_database=True):
             rows = (
                 session.query(
                     VarsPerHistory.history_id.label("history_id"),
-                    func.array_agg(VarsPerHistory.vars_id).label("variable_ids"),
+                    func.array_agg(VarsPerHistory.vars_id).label(
+                        "variable_ids"
+                    ),
                 )
                 .group_by(VarsPerHistory.history_id)
                 .all()
             )
-            return {
-                row.history_id: row.variable_ids for row in rows
-            }
+            return {row.history_id: row.variable_ids for row in rows}
 
     with log_timing("Query all vars by hx", log=logger.debug):
         all_variables = (
@@ -116,10 +79,10 @@ def get_all_vars_by_hx(session, group_in_database=True):
                 VarsPerHistory.history_id.label("history_id"),
                 VarsPerHistory.vars_id.label("id"),
             )
+            # TODO: Is ordering necessary? Probably not.
             .order_by(
                 VarsPerHistory.history_id.asc(), VarsPerHistory.vars_id.asc()
-            )
-            .all()
+            ).all()
         )
     with log_timing("Group all vars by hx", log=logger.debug):
         result = {
